% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_surveys.R
\name{browse_team_templates}
\alias{browse_team_templates}
\title{browse_team_templates}
\usage{
browse_team_templates(oauth_token = get_token())
}
\arguments{
\item{oauth_token}{Your OAuth 2.0 token. By default, retrieved from \code{get_token()}.}
}
\value{
A dataframe of team survey templates and associated metadata from desired account.
}
\description{
Retrieve a dataframe of survey templates available to the user's team.
}
\details{
This function calls the SurveyMonkey API using the current OAuth token and returns
a dataframe of survey template available to the user's team.
}
\references{
SurveyMonkey API V3 at \url{https://developer.surveymonkey.com/api/v3/#api-endpoints-get-team_survey_templates}.
}
