% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_survey.R
\name{code_educ_vars}
\alias{code_educ_vars}
\title{Code demographic variables in responses}
\usage{
code_educ_vars(data, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame of survey responses.}

\item{verbose}{Print helpful messages.}
}
\value{
Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' \bold{educational attainment}: \itemize{
 \item \code{educ2}: binary indicator for college vs. not
 \item \code{educ4}: 4-point educational attainment scale
 \item \code{educ}: 6-point educational attainment scale
}
}
\description{
Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use \code{code_std_demo_vars()} to perform all standard demographic recodes, or select the specific function you need.
}
