% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_survey.R
\name{code_ideo_vars}
\alias{code_ideo_vars}
\title{Code demographic variables in responses}
\usage{
code_ideo_vars(data, append = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame of survey responses.}

\item{append}{If TRUE, append to existing dataframe otherwise return new variables in a fresh dataframe.}

\item{verbose}{Print helpful messages.}
}
\value{
Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' \bold{political ideology}: \itemize{
 \item \code{ideo5}: 5-point political ideology (i.e. Very Cons, Cons, Mod, Lib, Very Lib)
 \item \code{ideo3}: 3-point political ideology (i.e. Cons, Mod, Lib)
}
}
\description{
Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use \code{code_std_demo_vars()} to perform all standard demographic recodes, or select the specific function you need. Run \code{std_demo_vars()} to return the full index of SurveyMonkey's standard demographic variables.
}
