% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_survey.R
\name{code_income_vars}
\alias{code_income_vars}
\title{Code demographic variables in responses}
\usage{
code_income_vars(data, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame of survey responses.}

\item{verbose}{Print helpful messages.}
}
\value{
Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' \bold{income levels}: \itemize{
 \item \code{income3}: 3-point income scale (i.e. <$50k, $50-100k, >$100k)
}
}
\description{
Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use \code{code_std_demo_vars()} to perform all standard demographic recodes, or select the specific function you need.
}
