% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_survey.R
\name{code_party_vars}
\alias{code_party_vars}
\title{Code demographic variables in responses}
\usage{
code_party_vars(data, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame of survey responses.}

\item{verbose}{Print helpful messages.}
}
\value{
Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' \bold{U.S. political party affiliation}: \itemize{
 \item \code{party5}: 5-point U.S. political party leaning (i.e. Dem, Lean Dem, Ind, Lean Rep, Rep)
 \item \code{party3}: 3-point U.S. political party affiliation (i.e. "forced" to Dem, Ind, Rep)
}
}
\description{
Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use \code{code_std_demo_vars()} to perform all standard demographic recodes, or select the specific function you need.
}
