% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_survey.R
\name{code_sex_vars}
\alias{code_sex_vars}
\title{Code demographic variables in responses}
\usage{
code_sex_vars(data, append = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame of survey responses.}

\item{append}{If TRUE, append to existing dataframe otherwise return new variables in a fresh dataframe.}

\item{verbose}{Print helpful messages.}
}
\value{
Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' \bold{gender identity and sexual orientation}: \itemize{
 \item \code{lgbtq}: 5-point sexual orientation scale (Lesbian, Gay, Bisexual, Heterosexual, Other)
 \item \code{lgbtq2}: 2-point sexual orientation scale (Heterosexual, Non-heterosexual)
 \item \code{transgender}: binary indicator for transgender identity or not
 \item \code{gender_binary}: binary indicator for gender
 \item \code{gender}: 3-point scale for gender allowing for non-binary gender identity
}
}
\description{
Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use \code{code_std_demo_vars()} to perform all standard demographic recodes, or select the specific function you need. Run \code{std_demo_vars()} to return the full index of SurveyMonkey's standard demographic variables.
}
