% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{design_eff}
\alias{design_eff}
\title{Estimate the Kish design effect of using a set of weights}
\usage{
design_eff(weights, na.rm = F)
}
\arguments{
\item{weights}{vector of sample weights used to adjust the sample.}

\item{na.rm}{trim entries that are NA.}
}
\value{
the ratio of the actual sample size to the effective sample size with weights.
}
\description{
The Kish design effect is how much the variance of a sample estimator 
is expected to increase as a result of post-hoc sample weights.
}
