% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ev22}
\alias{ev22}
\title{SVMK Electric Vehicles Survey 2022 Study}
\format{
A simulated data frame of partially coded survey responses with 1000 rows and 31 columns:
\describe{
  \item{gender}{Gender (standard research team format)}
  \item{age}{Age (standard research team format)}
  \item{education}{Education (standard research team format)}
  \item{incomse}{Income (standard research team format)}
  \item{vehage}{"How old is your main vehicle?"}
  \item{vehcost}{"In total, how much did you pay to buy or lease your main vehicle?"}
  \item{purchase}{"How did you purchase or lease your main vehicle?"}
  \item{ev_heard_*}{"Which of the following electric vehicles have you heard of?" (select all that apply)}
  \item{mech_factors_*}{"Select the top 3 factors that are most important to you when selecting an auto repair provider:" (select all that apply)}
  \item{sub_services_*}{"Which of the following services would you be willing to pay a recurring subscription for in a car?"  (select all that apply)}
  ...
}
}
\usage{
ev22
}
\description{
Simulated data based on a real market research study on Electric Vehicles conducted by the SurveyMonkey Research Insights Team.
}
\keyword{datasets}
