% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{fetch_survey_obj}
\alias{fetch_survey_obj}
\title{Download a survey's responses}
\usage{
fetch_survey_obj(id, oauth_token = get_token())
}
\arguments{
\item{id}{ID number of survey to be fetched.}

\item{oauth_token}{Your OAuth 2.0 token.
By default, retrieved from \code{get_token()}.}
}
\value{
A survey object of type \code{qdoc}, which is a nested list containing info about the survey and underlying questionnaire. Note that you can print this object for an informative summary like you can any other qdoc object.
}
\description{
Download a survey's responses
}
\examples{
# not run:
# fetch_survey_obj(123456789)
}
