% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_rules.R
\name{find_skip_rules}
\alias{find_skip_rules}
\title{Find and code skip rules from survey responses}
\usage{
find_skip_rules(
  data,
  interactive = TRUE,
  skip_rule_val = NA,
  no_answer_val = "No answer"
)
}
\arguments{
\item{data}{input survey data frame of responses (note: function only guaranteed to work if this is the output from \code{parse_survey}).}

\item{interactive}{if set to \code{TRUE}, detect possible skip rules and ask before re-coding responses; else automatically detect rules and re-code responses.}

\item{skip_rule_val}{what to code question responses where the respondent never saw the question due to a skip rule (default is \code{NA} which tends to play nicely with banner and cross-tab creation).}

\item{no_answer_val}{what to code question responses where the respondent saw the question and did not answer (default is "No answer").
By default, retrieved from \code{get_token()}.}
}
\value{
A survey response dataframe with NAs distinguished between "Not applicable" responses from "No answer" responses.
}
\description{
Given a dataframe of survey responses, reverse engineer skip rules based on completion patterns (and optionally user input) in order to distinguish/re-code "Not applicable" \code{NA} values from "No answer" \code{NA} values.
}
\details{
Note: if \code{skip_rule_val} or \code{no_answer_val} are set to character strings, the corresponding factor variables (in the relevant question columns) will be expanded via the \code{forcats} library to include those values as levels.
}
\examples{
# TODO
}
