% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_surveys.R
\name{get_target}
\alias{get_target}
\title{Get a SurveyMonkey target population}
\usage{
get_target(name)
}
\arguments{
\item{id}{name of target population to read in (see \code{list_targets} for options).}
}
\value{
list of dataframes associated with named target population profile.
}
\description{
Read in a default target population (i.e. specific variables and their marginal and joint distributions) available through this package for you to weight your survey to.
}
