% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_surveys.R
\name{list_targets}
\alias{list_targets}
\title{Look up SurveyMonkey target populations}
\usage{
list_targets()
}
\description{
Look up all of the target population profiles (i.e. specific variables and their marginal and joint distributions) available to weight your survey to.
}
