% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{n_eff}
\alias{n_eff}
\title{Estimate the Kish effective sample size of using a set of weights.}
\usage{
n_eff(weights, na.rm = F)
}
\arguments{
\item{weights}{vector of sample weights used to adjust the sample.}

\item{na.rm}{trim entries that are NA.}
}
\value{
the sample size that produces equivalent variance when introducing the sample weights.
}
\description{
The Kish effective sample size is the reduced size of our sample given the increase in variance.
}
