% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{parse_survey}
\alias{parse_survey}
\title{Take a survey object and parses it into a tidy data.frame (optimized).}
\usage{
parse_survey(surv_obj, oauth_token = get_token(), col_names = "id", ...)
}
\arguments{
\item{surv_obj}{a survey, the result of a call to \code{fetch_survey_obj}.}

\item{oauth_token}{Your OAuth 2.0 token. By default, retrieved from
\code{get_token()}.}

\item{col_names}{set column names either to question IDs or names. Options: "id" or "name". Default: "id".}

\item{...}{additional arguments to pass on to \code{get_responses}.  See the documentation
\code{?get_responses} where these arguments are listed.}
}
\value{
a data.frame (technically a \code{tibble}) with clean responses, one line per respondent.
}
\description{
Take a survey object and parses it into a tidy data.frame (optimized).
}
