% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_survey.R
\name{qdoc.question}
\alias{qdoc.question}
\alias{as.qdoc.question}
\alias{qdoc.q}
\title{Process a questionnaire question}
\usage{
qdoc.question(text, quiet.fail = FALSE)

as.qdoc.question(text, quiet.fail = FALSE)

qdoc.q(text, quiet.fail = FALSE)
}
\arguments{
\item{text}{text of question written in proper QDOC format.}

\item{quiet.fail}{if cannot parse question, silently return an empty list (default is FALSE).}
}
\value{
An object for a particular QDOC question family (e.g. \code{qdoc.single_choice}, \code{qdoc.multiple_choice}, \code{qdoc.matrix}).
}
\description{
Takes a question written in proper QDOC format and 
generates an object that is ready to be uploaded as a survey.
See \code{vignette("qdocs")} for a guide to the QDOC format.
}
\examples{
# multiple-choice (single answer)
qdoc.question("What is your gender?\n\nMale\nFemale\nOther")
qdoc.question("[[Question:MC]]\nWhat is your gender?[[Choices]]\nMale\nFemale\nOther")

# multiple-choice (multi-answer)
qdoc.question("What is your gender?[[MultiAnswer]]\n\nMale\nFemale\nOther")
qdoc.question("[[Question:MC:MultiAnswer]]\nWhat is your gender?\n[[Choices]]\nMale\nFemale\nOther")

# matrix
qdoc.question("Please rate the following:\n\na\nb\nc\n\n1\n2\n3")
qdoc.question("[[Question:Matrix]]\nPlease rate the following:\n[[Rows]]\na\nb\nc\n[[Columns]]\n1\n2\n3")
}
