% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_survey.R
\name{read_qdoc}
\alias{read_qdoc}
\title{Process a questionnaire}
\usage{
read_qdoc(file = NULL, gdoc = NULL, text = NULL, oauth_token = get_token())
}
\arguments{
\item{file}{File for questionnaire written in proper QDOC format.}

\item{gdoc}{File in Google Drive for Google Doc questionnaire written in proper QDOC format. See \code{drive_download} for acceptable formats (e.g. URL, file ID, or \code{dribble} object if you're fancy).}

\item{text}{Character string written in in proper QDOC format.}

\item{oauth_token}{Your OAuth 2.0 token (only needed if questionnaire references question bank or templates). By default, retrieved from get_token().}
}
\value{
Output of class \code{qdoc}.
}
\description{
Takes a questionnaire written in proper QDOC format and 
generates an object that is ready to be uploaded as a survey.
See \code{vignette("qdocs")} for a guide to the QDOC format.
}
\examples{
data(qdocs)

set_token("XXXXXXX")

read_qdoc(text = qdoc.simple) # simple format
read_qdoc(text = qdoc.adv)    # advanced format

}
