% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{simu_moe}
\alias{simu_moe}
\title{Simulate the Margin of Error (or "modelled error") for a survey.}
\usage{
simu_moe(weights, conf = 0.95, sims = 5000, random = FALSE)
}
\arguments{
\item{weights}{vector of sample weights used to adjust the sample.}

\item{conf}{the level of statistical confidence to estimate the error.}

\item{sims}{number of bootstrap re-samples to perform in order to simulate the sampling distribution.}

\item{random}{whether or not to return different, random estimates every time.}
}
\value{
the survey margin of error (SMOE) or "modeled error" for the overall survey.
}
\description{
Given weights and a sample size, the Modeled Error gives us the upper bound on the margin of error
for all y/n questions across the entire survey -- by estimating it
for a hypothetical y/n question with 50/50% split in the population (the inherently
most uncertain population estimand).
}
\details{
Estimates via simulation tend to be more conservative (larger) than 
using the formulaic margin of error (\code{est_modeled_error}) as it incorporates 
more variance from the weights. 

This method is recommended for non-probability surveys with unknown sampling mechanisms (e.g. river samples).
}
