% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_utils.R
\name{sm_post}
\alias{sm_post}
\title{POST request for svmkR API}
\usage{
sm_post(url, body, config, ...)
}
\arguments{
\item{url}{character, url to send request to}

\item{body}{character, request body for POST call.}

\item{config}{see \code{?httr::config}, \link[httr]{config}, for full details.
Additional configuration settings such as http additional headers.}

\item{...}{additional argument passed to \code{httr::POST} \link[httr]{POST}.}
}
\description{
Will always set the \code{soubhikbarari/svmkR} repo as the user agent
}
