% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smda23}
\alias{smda23}
\title{SVMK Social Media Usage 2023 Study}
\format{
A simulated data frame of raw survey responses with 4000 rows and 81 columns with the following identifying variables for each respondent (mostly redacted):
\describe{
  \item{collector_id}{Survey collector ID (redacted)}
  \item{collection_mode}{Survey collection mode}
  \item{survey_id}{Survey ID (redacted)}
  \item{response_id}{Unique respondent ID (redacted)}
  \item{response_status}{Response staus}
  \item{date_created}{Date response started}
  \item{date_modified}{Date response last modified}
  \item{first_name}{Respondent's first name (redacted)}
  \item{last_name}{Respondent's last name (redacted)}
  \item{email_address}{Respondent's email address (redacted)}
  \item{ip_address}{Respondent's IP address (redacted)}
  ...
}
}
\usage{
smda23
}
\description{
Simulated data based on a real tracker study conducted by the SurveyMonkey Research Insights Team on social media usage and other behaviors/attitudes.
}
\keyword{datasets}
