% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_surveys.R
\name{weight_to}
\alias{weight_to}
\title{Weight your survey}
\usage{
weight_to(
  data,
  target = "us_genpop_acs18",
  auto.remove = TRUE,
  initial.weights = NULL,
  trim.weights = c(0.01, 0.99),
  verbose = TRUE
)
}
\arguments{
\item{data}{input survey data frame.}

\item{target}{either name of target population (run \code{list_targets()} for options) or a list of dataframes corresponding to different population distributions to weight survey to.}

\item{auto.remove}{remove any weighting variables that can't be found in survey.}

\item{initial.weights}{initial set of weights as starting point for raking algorithm.}

\item{trim.weights}{percentiles to trim your weights (default are 0.01 and .99, or 1\% and 99\%); can specify either an upper percentile or both a lower and upper percentile.}

\item{verbose}{output helpful progress and warning messages (recommended).}
}
\description{
Generate statistical weights for your survey responses to better represent some population of interest.
}
\details{
The default SurveyMonkey to weight surveys is via raking. For an overview of raking, see \href{https://www.abtassociates.com/raking-survey-data-aka-sample-balancing}{this guide}.
}
