% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_surveys.R
\name{weight_to_US_genpop}
\alias{weight_to_US_genpop}
\title{Weight your survey}
\usage{
weight_to_US_genpop(
  data,
  target = "acs18",
  remove.vars = TRUE,
  initial.weights = NULL,
  trim.weights = c(0.01, 0.99)
)
}
\arguments{
\item{data}{input survey data frame.}

\item{target}{name of target population, see \code{us_genpop_targets()} for options.}

\item{remove.vars}{remove any weighting variables that can't be found in survey.}

\item{initial.weights}{initial set of weights as starting point for raking algorithm.}

\item{trim.weights}{percentiles to trim your weights (default are 0.01 and .99, or 1\% and 99\%); can specify either an upper percentile or both a lower and upper percentile.}
}
\description{
Generate statistical weights for your survey responses to better represent some population of interest.
}
\details{
The default SurveyMonkey to weight surveys is via raking. For an overview of raking, see \href{https://www.abtassociates.com/raking-survey-data-aka-sample-balancing}{this guide}.
}
