% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banners.R
\name{write_banners}
\alias{write_banners}
\title{Format and save banner object to file / Google Drive.}
\usage{
write_banners(
  banners.output,
  title = "Main",
  file.path,
  file.overwrite = TRUE,
  file.append = FALSE,
  drive.folder.path = NULL,
  drive.overwrite = TRUE,
  tab.name = "Main",
  logo = "mntv",
  include.moe = TRUE,
  include.dates = TRUE,
  date.var = "StartDate",
  weight.var = NULL,
  ...
)
}
\arguments{
\item{banners.output}{output list from a call to \code{make_banners()}.}

\item{title}{title of the tab/sheet.}

\item{file.path}{local file path to save banners as \code{.xlsx} file.}

\item{file.overwrite}{whether or not to overwrite the file specified at \code{file.path}; otherwise keep both and update the file path.}

\item{file.append}{whether or not to append the current tab to the file specified at \code{file.path} if it exists; if there is already a tab with the name specified, update the tab name.}

\item{drive.folder.path}{folder path in Google Drive to upload banners spreadsheet.}

\item{drive.overwrite}{if \code{drive.folder.path} is specified, whether or not to overwrite/update file in Google Drive if it already exists; otherwise upload the file with the same name but an updated timestamp.}

\item{tab.name}{name of the tab/sheet to populate.}

\item{logo}{whether or not to include either the SurveyMonkey logo ("svmk") or the Momentive logo ("mntv") in the banner headers.}

\item{include.moe}{whether or not to include the Margin of Error for the survey in the banner headers.}

\item{include.dates}{whether or not to include the date range for survey responses in the banner headers.}

\item{date.var}{if \code{include.dates} is set to \code{TRUE}, the date variable to look for in the data used to create \code{banners.output}.}

\item{weight.var}{the column in the data used to create \code{banners.output} corresponding to weights for the Margin of Error estimate.}
}
\value{
a fully formatted spreadsheet in the form of an \code{openxlsx} object.
}
\description{
Take the output from a call to \code{make_banners}, populate a spreadsheet with the banner data, format/stylize it in the SurveyMonkey research format, and save it locally and/or to a Google Drive folder.
}
