% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_survey.R
\name{code_geo_vars}
\alias{code_geo_vars}
\title{Code demographic variables in responses}
\usage{
code_geo_vars(data, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame of survey responses.}

\item{verbose}{Print helpful messages.}
}
\value{
Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' \bold{geographic location (currently only in the U.S.)}: \itemize{
 \item \code{zip}: 6-digit ZIP code
 \item \code{state}: state name
 \item \code{state_abb}: state abbreviation
 \item \code{state_fips}: state FIPS code
 \item \code{region}: U.S. Census region
 \item \code{division}: U.S. Census division
}
}
\description{
Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use \code{code_std_demo_vars()} to perform all standard demographic recodes, or select the specific function you need.
}
