% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{esti_moe}
\alias{esti_moe}
\title{Estimate the Margin of Error (or "modelled error") for a survey.}
\usage{
esti_moe(weights = NULL, conf = 0.95)
}
\arguments{
\item{weights}{vector of sample weights used to adjust the sample.}

\item{conf}{the level of statistical confidence to estimate the error.}
}
\value{
the survey margin of error (SMOE) or "modeled error" for the overall survey.
}
\description{
Given weights and a sample size, the Modeled Error gives us the upper bound on the margin of error
for all y/n questions across the entire survey -- by estimating it
for a hypothetical y/n question with 50/50% split in the population (the inherently
most uncertain population estimand).
}
\details{
Estimates from this asymptotic formula tend to be smaller than
using the simulated margin of error (\code{sim_modeled_error}) as it assumes 
all our survey's questions follow an asymptotic normal distribution 
(i.e. a large sample of independent observations). 

For non-probability surveys without design weights, it is recommended to use `sim_modeled_error`.
}
