% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_surveys.R
\name{get_target}
\alias{get_target}
\title{Get a SurveyMonkey target population}
\usage{
get_target(name)
}
\description{
Read in a default target population  (i.e. specific variables and their marginal and joint distributions) available through this package for you to weight your survey to.
}
