% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_survey.R
\name{parse_survey}
\alias{parse_survey}
\title{Take a survey object and parse it into a tidy dataframe.}
\usage{
parse_survey(
  surv_obj,
  oauth_token = get_token(),
  col_names = "id",
  col_fill = TRUE,
  ...
)
}
\arguments{
\item{surv_obj}{A survey, the result of a call to \code{fetch_survey_obj}.}

\item{oauth_token}{Your OAuth 2.0 token. By default, retrieved from
\code{get_token()}.}

\item{col_names}{Set column names either to names ("name") or IDs ("id", default). Depending on question types, column names will be formatted as one of the following where each `{}` element is either an ID or a name:
\itemize{
  \item '\{question heading\}'
  \item '\{question heading\} - Other'
  \item '\{question heading\} - \{matrix row\}'
  \item '\{question heading\} - \{matrix row\} - \{matrix column\}'
}}

\item{col_fill}{For the columns extracted from a multi-select question, how to fill in values. Choices are either TRUE or "name" denoting the name of each question choice. Default is TRUE.}

\item{...}{Additional arguments to pass on to \code{get_responses}.  See the documentation
\code{?get_responses} where these arguments are listed.}
}
\value{
A data.frame (technically a \code{tibble}) with clean responses, one line per respondent.
}
\description{
Take a survey object and parse it into a tidy dataframe.
}
