% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{simu_moe}
\alias{simu_moe}
\title{Simulate the margin of error for an overall survey or a single question}
\usage{
simu_moe(
  .data = NULL,
  weights = NULL,
  n = NULL,
  x = NULL,
  conf = 0.95,
  sims = 5000,
  random = FALSE
)
}
\arguments{
\item{.data}{data frame to optionally pipe in \code{dplyr}-style (see example).}

\item{weights}{vector of sample weights used to adjust the sample (optional if \code{n} or \code{x} is specified).}

\item{n}{the sample size of the suvrey (not needed if \code{weights} or \code{x} is specified).}

\item{x}{the particular question to estimate margin of error for (if not specified, estimate the survey's MOE instead).}

\item{conf}{the level of statistical confidence to estimate the error.}

\item{sims}{number of bootstrap re-samples to perform in order to simulate the sampling distribution.}

\item{random}{whether or not to return different, random estimates every time.}
}
\value{
The survey margin of error (MOE) or "modeled error" for the overall survey or question of interest.
}
\description{
For the overall survey: given weights and/or a sample size, the 
maximum margin of error gives us the upper bound on the margin of error for all y/n questions across 
the entire survey -- by estimating it for a hypothetical y/n question with 50/50% split in the population 
(the inherently most uncertain population estimand).
}
\details{
For a single question: given a vector of sample responses and/or weights, the margin of error
estimates the margin of error by simulating the sampling distribution.

Estimates via simulation tend to be more conservative (larger) than 
using the formulaic margin of error (\code{est_modeled_error}) as it incorporates 
more variance from the weights. 

This method is recommended for non-probability surveys with unknown sampling mechanisms (e.g. river samples).
}
\examples{

data(ev22)
if (FALSE) { ## not run

# Unweighted maximum margin of error
simu_moe(n = nrow(ev22))
simu_moe(weights = rep(1, nrow(ev22)))

# Weighted maximum margin of error
simu_moe(weights = ev22$weight_genpop)

# Unweighted margin of error for specific question
simu_moe(x = ev22$ev_heard_1)

# Weighted margin of error for specific question
ev22 \%>\%
  filter(!is.na(weight_genpop)) \%>\%
  simu_moe(x = ev_heard_1, weights = weight_genpop)
}
}
